refactor(domain): firecrawlモジュールをservicesディレクトリに移動
firecrawl関連のコードをdomain直下からdomain/services配下に移動
これによりサービス層のコードを一箇所に集約し、構造を明確化
