+++ src/domain/mod.rs
@@ -1,4 +1,4 @@
 pub mod article;
 pub mod feed;
-pub mod firecrawl;
 pub mod rss;
+pub mod services;
