+++ src/domain/services/mod.rs
@@ -1,0 +1,1 @@
+pub mod firecrawl;
